refactor: move all branch ops into one file
