--- src/lib/git/branch_move.ts
@@ -1,9 +1,0 @@
-import { q } from '../utils/escape_for_shell';
-import { gpExecSync } from '../utils/exec_sync';
-
-export function branchMove(newName: string): void {
-  gpExecSync({
-    command: `git branch -m ${q(newName)}`,
-    onError: 'throw',
-  });
-}
