--- src/lib/git/current_branch_name.ts
@@ -1,10 +1,0 @@
-import { gpExecSync } from '../utils/exec_sync';
-
-export function getCurrentBranchName(): string | undefined {
-  const branchName = gpExecSync({
-    command: `git branch --show-current`,
-    onError: 'ignore',
-  });
-
-  return branchName.length > 0 ? branchName : undefined;
-}
